+++ src/test/groovy/ru/vyarus/spock/jupiter/test/exception/SpockTestExceptionHandler3.groovy
@@ -4,18 +4,17 @@ import org.junit.jupiter.api.extension.ExtendWith
 import ru.vyarus.spock.jupiter.AbstractTest
 import ru.vyarus.spock.jupiter.support.LifecycleExtension
 import ru.vyarus.spock.jupiter.support.RethrowExceptionHandler
-import ru.vyarus.spock.jupiter.support.EatExceptionHandler
 import spock.lang.Requires
 import spock.lang.Specification
 
 /**
  * @author Vyacheslav Rusakov
- * @since 02.01.2022
+ * @since 03.02.2026
  */
 @Requires({ AbstractTest.ACTIVE })
-// exception handlers processing order is reversed
-@ExtendWith([LifecycleExtension, EatExceptionHandler, RethrowExceptionHandler])
-class SpockTestExceptionHandler2 extends Specification {
+// exception handlers processing order is reversed, exception hot handled (rethrowed)
+@ExtendWith([LifecycleExtension, RethrowExceptionHandler])
+class SpockTestExceptionHandler3 extends Specification {
 
     def "Sample test"() {