+++ src/test/java/playground/tests/exceptions/JupiterTestExceptionHandler3.java
@@ -1,23 +1,23 @@
 package playground.tests.exceptions;
 
+import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import ru.vyarus.spock.jupiter.support.LifecycleExtension;
 import ru.vyarus.spock.jupiter.support.RethrowExceptionHandler;
-import ru.vyarus.spock.jupiter.support.EatExceptionHandler;
 
 /**
  * @author Vyacheslav Rusakov
- * @since 02.01.2022
+ * @since 03.02.2026
  */
 @Disabled
-// exception handlers processing order is reversed
-@ExtendWith({LifecycleExtension.class, EatExceptionHandler.class, RethrowExceptionHandler.class})
-public class JupiterTestExceptionHandler {
+// exception handlers processing order is reversed, exception hot handled (rethrowed)
+@ExtendWith({LifecycleExtension.class, RethrowExceptionHandler.class})
+public class JupiterTestExceptionHandler3 {
 
     @Test
     void sampleTest() {
-        throw new IllegalStateException("problem");
+        Assertions.fail("assert fail");
     }
 }