+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 * Update junit compatibility
     - [junit 5.11](https://docs.junit.org/5.11.0/release-notes/#release-notes-5.11.0-junit-jupiter-deprecations-and-breaking-changes): 
-      @ExtendWith on non-static field now participate in beforeAll/afterAll;
-      static extension fields are registered before non-static fields (fixed order)
+      * @ExtendWith on a non-static field now participate in beforeAll/afterAll;
+      * static extension fields are registered before non-static fields (junit behavior change)
+* Add support for LifecycleMethodExecutionExceptionHandler (for spock's setupSpec, setup etc.)
 
 ### 1.2.0 (2022-11-25)
 * Add root (engine-level) context to be able to use global storage in extensions (#44)